+++ changelog.md
@@ -1,3 +1,8 @@
-# 1.0.0
+# 1.1.2
+
+- Added ability to toggle practice with the key "C".
+- Removed notification feature, it was useless and annoying.
+
+# 1.1.1
 
 - Initial Release.